Split graph-preview.tsx: 2,297 -> 1,097 lines, five modules
First of the three files named in #281. Two kinds of extraction, and the
issue's own guidance was to prefer the first:

TESTABLE LOGIC -> .ts, which is the real leverage. This app's vitest cannot
parse .tsx at all, so every line of this was unreachable by a unit test purely
because of where it lived:

  preview-time-channel.ts   (155)  time/scrub/play/audio state + persistence
  preview-card-geometry.ts   (90)  card widths, and which cards an overlay measures

15 new tests, all first coverage. Nothing about them is new behaviour.

SHARED SEAMS -> their own modules, because the seek rails needed them and
re-exporting from graph-preview would have made the two files import each
other. A cycle around module-level createContext calls fails at evaluation
time, not at the type level, so this was worth avoiding rather than working
around:

  preview-contexts.tsx       (88)  PreviewCardSpans + FlatItems contexts
  preview-scroller.ts        (75)  scroller lookup + ruler tick window

RELOCATION, no behaviour change:

  graph-seek-rails.tsx      (965)  the grid's per-row sliders and the strip rail

graph-preview.tsx re-exports everything that moved, so no consumer's import
path changed — graph-board.tsx and graph-sub-timelines.tsx are untouched.

Verified beyond the type checker, because ~965 lines of moved JSX have no unit
coverage: 145/145 graph-view e2e, and the seek rail driven by keyboard in the
running app (0.0 -> 5.0s over five arrow presses, 6 rails mounted with correct
bounds).
